Allow textures to be moved from one GL context group to another.
The function glProduceTexture(target, mailbox) removes a texture from a context group and puts it in a "mailbox" that is accessible to all context groups on the same channel.

Then glConsumeTexture(target, mailbox) adds the texture to a context group, possibly a different one, if the caller knows the name of the mailbox.

From the point of view of the caller, the texture objects themselves do not move, just the texels. In terms of the underlying GL, the TextureInfo service IDs are reassigned as necessary.

A texture in a mailbox is destroyed when the context group that put it there is destroyed.

Expected usage is:

// Context 1
glBindTexture(GL_TEXTURE_2D, foo);
glProduceTexture(GL_TEXTURE_2D, secret_name);
glFlush();

// Context 2
glBindTexture(GL_TEXTURE_2D, bar);
glConsumeTexture(GL_TEXTURE_2D, secret_name);

Review URL: http://codereview.chromium.org/10106015

git-svn-id: http://src.chromium.org/svn/trunk/src/third_party/khronos@134791 4ff67af0-8c30-449e-8e8b-ad334ec8d88c