Addition of an unpremultiplied-alpha setting for the GL_CHROMIUM_copy_texture command buffer GL extension.
The extension needs to support the ability to unpremultiply the source of the copy because of the canvas2d-webgl texture use case.  Canvas is stored in premultiplied form, yet WebGL expects texture contents to be in unpremultiplied form.

Change also includes conversion of argument types to overriders of WebGraphicsContext3D::copyTextureCHROMIUM(...).  
This change is necessary as a result of a wk-change:  https://bugs.webkit.org/show_bug.cgi?id=86275

BUG=101051
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10535174

git-svn-id: http://src.chromium.org/svn/trunk/src/third_party/khronos@142962 4ff67af0-8c30-449e-8e8b-ad334ec8d88c