Add a function to provide the list of third-party directories which are incompatible with Android.
For most directories, we use a regex of whitelisted licenses to parse the 'License' field from the metadata. In some case, for example, where a custom license is used, we rely on a new 'Android Compatibility' field.

This functionality will be used by the snapshot tool, which will exclude these incompatible directories from the snapshot in the Android tree.

BUG=138921

Review URL: https://chromiumcodereview.appspot.com/10829272

git-svn-id: http://src.chromium.org/svn/trunk/src/third_party/khronos@151462 4ff67af0-8c30-449e-8e8b-ad334ec8d88c